Struct configure::source::ActiveConfiguration

pub struct ActiveConfiguration { /* fields omitted */ }

The active configuration source.

The only value of this type is the CONFIGURATION global static, which controls what the source of configuration values is. End users can set the configuration source using the set method, while libraries which need to be configured can use the get method.

Methods

impl ActiveConfiguration

pub fn set<T: ConfigSource>(&'static self, source: T)

Set the active configuration.

This can only be called once. If it is called more than once, subsequent calls have no effect. This should only be called by the final binary which is using the configuration, it should not be called by libraries.

If you set the active configuration, you should do so very early in your program, preferably as close to the beginning of main as possible. That way, the configuration source is consistent for every dependency.

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get(&'static self, package: &'static str) -> Box<DynamicDeserializer>

Get the active configuration.

Libraries which need to construct configuration can use this to get the active source of configuration. Normally they would derive Configure for their config struct, which will call this method.

pub fn is_default(&'static self) -> bool

Returns true if the configuration source is the default source.

The opposite of CONFIGURATION.is_overriden()

pub fn is_overriden(&'static self) -> bool

Returns true if the configuration source has been overriden.

The opposite of CONFIGURATION.is_default()