Struct configure::Opt

struct Opt {
    no_testlib: bool,
    no_man: bool,
    no_tools: bool,
    no_core: bool,
    extras: bool,
    archive: bool,
    cargo: Option<String>,
    cmake: Option<String>,
    target: Option<String>,
    optimize: bool,
    dbg_sym: bool,
    verbose: bool,
    out_dir: Option<PathBuf>,
    deb: bool,
    deb_opt: Vec<String>,
    systemd: bool,
    install_prefix: Option<PathBuf>,
    docker: bool,
    docker_tag: Vec<String>,
    docker_build_opt: Vec<String>,
    apidoc: bool,
    docker_name: Option<String>,
    json_schema: bool,
    compile_trial_contest: bool,
    rust_doc: bool,
}

Fields
no_testlib: bool

Do not build and install testlib

no_man: bool

Do not generate and install manual

no_tools: bool

Do not build and install additional tools

no_core: bool

Do not build and install JJS core components

extras: bool

Build and install extras (components, not directly related to JJS)

archive: bool

Generate tarball

cargo: Option<String>

Cargo path

cmake: Option<String>

CMake path

target: Option<String>

Target triple

optimize: bool

Optimization

dbg_sym: bool

Debug symbols

verbose: bool

Emit verbose information about build

out_dir: Option<PathBuf>

Destination for artifacts

deb: bool

Build deb packages

deb_opt: Vec<String>

Pass given option to deb generator

systemd: bool

Generate SystemD unit files

install_prefix: Option<PathBuf>

Destination JJS will be installed to

Some JJS components can not be built or work properly without this option By default, same as prefix.

docker: bool

Build docker images

docker_tag: Vec<String>

Docker image tag

docker_build_opt: Vec<String>

Docker build additional options

apidoc: bool

Build apiserver API documentation. Requires npx

docker_name: Option<String>

Name or path to Docker or other tool which can run containers (e.g. Podman)

json_schema: bool

If enabled, json schemas will be emitted

compile_trial_contest: bool

If enabled, trial contest problems will be compiled and packaged into archive

rust_doc: bool

If enabled, Rust API documentation will be generated for all crates

Methods

impl Opt

fn process_deps(&mut self)